Update route-service to adapt to the new version of Neo4j database: cityId as the business primary key, distanceKm as double

+++ src/main/java/fr/dauphine/miageIf/minh/yang/route_service/dto/CreationEdgeRequest.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 public class CreationEdgeRequest {
     private String sourceCityId;
     private String destinationCityId;
-    private int distanceKm;
+    private double distanceKm;
     private int travelTimeMin;
 }
